move FakeGTP to /utils/test
